Split large files and trim test class
